Initial circle ci setup
This CL adds a circle yml file to run tests of all build files.
It unfortunately uses v1 so it is very slow. Emulator may also
lock during tests, which i'm trying to mitigate in the runner
shell script but has no guaratee. I didn't want to change our
UI tests to unlock the device but we can do that as well.