Upgrade Navigation, AppCompat, Fragments, and Lifecycle
Upgrade to the latest releases:

- Navigation 2.2.0-alpha01
- AppCompat 1.1.0-rc01
- Fragment 1.2.0-alpha02
- Lifecycle 2.2.0-alpha03

As Navigation and the -ktx modules are built with Java 8 bytecode, many of the projects now also use Java 8.

Any times we were using a FrameLayout as a container for fragments, it has been switched to the new FragmentContainerView.

This change also involves replacing the deprecated ViewModelProviders.of() with either `by viewModels()` or the `ViewModelProvider` constructor and removing all dependencies on lifecycle-extensions, which only contains deprecated classes at this point.

Also took the opportunity to remove support.v4 from all projects as it was not being used anywhere.

Test: ./run_all_tests.sh