First commit for DonutTracker app, a sample for the MAD Skills Navigation series
This app consists of a list of items (donuts) in a RecyclerView
and a dialog used to enter or edit information about those items.
The sample is mainly useful in conjunction with the MAD Skills series
of articles and videos that use the sample; there are other samples
in this repo that are intended for more general-purpose exmaples
of the Navigation Component.